feat : 오류 수정 및 directory 이동
