+++ order-api/src/main/java/com/zerobase/cms/order/ZeroOrderApplication.java
@@ -1,23 +1,20 @@
-package com.zerobase.cms.user;
+package com.zerobase.cms.order;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.web.servlet.ServletComponentScan;
 import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.data.envers.repository.support.EnversRevisionRepositoryFactoryBean;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
-@EnableJpaRepositories
-@ServletComponentScan
-@EnableJpaAuditing
-@RequiredArgsConstructor
-@EnableSwagger2
 @SpringBootApplication
 @EnableFeignClients
-public class UserApplication {
+@EnableJpaAuditing
+@EnableJpaRepositories(repositoryFactoryBeanClass = EnversRevisionRepositoryFactoryBean.class)
+@ServletComponentScan
+public class ZeroOrderApplication {
     public static void main(String[] args) {
-        SpringApplication.run(UserApplication.class,args);
+        SpringApplication.run(ZeroOrderApplication.class,args);
     }
 }
