+++ order-api/src/main/java/com/zerobase/cms/order/config/JwtConfig.java
@@ -1,0 +1,15 @@
+package com.zerobase.cms.order.config;
+
+import com.zerobase.domain.config.JwtAuthenticationProvider;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class JwtConfig {
+
+    @Bean
+    public JwtAuthenticationProvider jwtAuthenticationProvider(){
+        return new JwtAuthenticationProvider();
+    }
+
+}
