+++ order-api/src/main/java/com/zerobase/cms/order/domain/product/ProductItemDto.java
@@ -1,0 +1,26 @@
+package com.zerobase.cms.order.domain.product;
+
+import com.zerobase.cms.order.domain.model.ProductItem;
+import lombok.*;
+
+@Getter
+@Builder
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProductItemDto {
+    private Long id;
+    private String name;
+    private Integer price;
+    private Integer count;
+
+    public static ProductItemDto from(ProductItem item){
+        return ProductItemDto.builder()
+                .id(item.getId())
+                .name(item.getName())
+                .price(item.getPrice())
+                .count(item.getCount())
+                .build();
+
+    }
+}
