+++ order-api/src/main/java/com/zerobase/cms/order/domain/product/UpdateProductForm.java
@@ -1,0 +1,20 @@
+package com.zerobase.cms.order.domain.product;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UpdateProductForm {
+    private Long id;
+    private String name;
+    private String description;
+    private List<UpdateProductItemForm> items;
+
+}
