+++ order-api/src/main/java/com/zerobase/cms/order/service/ProductSearchService.java
@@ -1,0 +1,9 @@
+package com.zerobase.cms.order.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ProductSearchService {
+}
