--- order-api/src/main/java/com/zerobase/cms/user/order/ZeroOrderApplication.java
@@ -1,20 +1,0 @@
-package com.zerobase.cms.user.order;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.servlet.ServletComponentScan;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.data.envers.repository.support.EnversRevisionRepositoryFactoryBean;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@SpringBootApplication
-@EnableFeignClients
-@EnableJpaAuditing
-@EnableJpaRepositories(repositoryFactoryBeanClass = EnversRevisionRepositoryFactoryBean.class)
-@ServletComponentScan
-public class ZeroOrderApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(ZeroOrderApplication.class,args);
-    }
-}
