--- order-api/src/main/java/com/zerobase/cms/user/order/config/JwtConfig.java
@@ -1,15 +1,0 @@
-package com.zerobase.cms.user.order.config;
-
-import com.zerobase.domain.config.JwtAuthenticationProvider;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class JwtConfig {
-
-    @Bean
-    public JwtAuthenticationProvider jwtAuthenticationProvider(){
-        return new JwtAuthenticationProvider();
-    }
-
-}
