--- order-api/src/main/java/com/zerobase/cms/user/order/domain/model/ProductItem.java
@@ -1,41 +1,0 @@
-package com.zerobase.cms.user.order.domain.model;
-
-import com.zerobase.cms.user.order.domain.product.AddProductItemForm;
-import lombok.*;
-import org.hibernate.envers.AuditOverride;
-import org.hibernate.envers.Audited;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@AuditOverride(forClass = BaseEntity.class)
-public class ProductItem extends BaseEntity{
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private Long sellerId;
-    @Audited
-    private String name;
-    @Audited
-    private Integer price;
-    private Integer count;
-
-    @ManyToOne
-    @JoinColumn(name = "product_id")
-    private Product product;
-
-    public static ProductItem of(Long sellerId, AddProductItemForm form){
-        return ProductItem.builder()
-                .sellerId(sellerId)
-                .name(form.getName())
-                .price(form.getPrice())
-                .count(form.getCount())
-                .build();
-    }
-}
