--- order-api/src/main/java/com/zerobase/cms/user/order/domain/product/AddProductItemForm.java
@@ -1,18 +1,0 @@
-package com.zerobase.cms.user.order.domain.product;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class AddProductItemForm {
-    private Long productId;
-    private String name;
-    private Integer price;
-    private Integer count;
-
-}
