--- order-api/src/main/java/com/zerobase/cms/user/order/domain/product/UpdateProductForm.java
@@ -1,20 +1,0 @@
-package com.zerobase.cms.user.order.domain.product;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class UpdateProductForm {
-    private Long id;
-    private String name;
-    private String description;
-    private List<UpdateProductItemForm> items;
-
-}
