--- order-api/src/main/java/com/zerobase/cms/user/order/domain/respository/ProductItemRepository.java
@@ -1,7 +1,0 @@
-package com.zerobase.cms.user.order.domain.respository;
-
-import com.zerobase.cms.user.order.domain.model.ProductItem;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductItemRepository extends JpaRepository<ProductItem,Long> {
-}
