--- order-api/src/main/java/com/zerobase/cms/user/order/domain/respository/ProductRepository.java
@@ -1,16 +1,0 @@
-package com.zerobase.cms.user.order.domain.respository;
-
-import com.zerobase.cms.user.order.domain.model.Product;
-import org.springframework.data.jpa.repository.EntityGraph;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface ProductRepository extends JpaRepository<Product,Long> {
-
-    @EntityGraph(attributePaths = {"productItems"},type = EntityGraph.EntityGraphType.LOAD)
-    Optional<Product> findBySellerIdAndId(Long sellerId,Long Id);
-
-    @EntityGraph(attributePaths = {"productItems"},type = EntityGraph.EntityGraphType.LOAD)
-    Optional<Product> findWithProductItemsById(Long id);
-}
