--- order-api/src/main/java/com/zerobase/cms/user/order/service/ProductSearchService.java
@@ -1,9 +1,0 @@
-package com.zerobase.cms.user.order.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class ProductSearchService {
-}
