+++ order-api/src/main/resources/application.yml
@@ -13,7 +13,7 @@ spring:
 
   jpa:
     show-sql: true
-    hibernate.ddl-auto: create
+    hibernate.ddl-auto: create-drop
 
   redis:
     host: localhost
