+++ user-api/build.gradle
@@ -25,7 +25,10 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'mysql:mysql-connector-java'
 
+
     runtimeOnly 'com.h2database:h2'
+
+    implementation 'org.apache.commons:commons-lang3:3.12.0'
 
     implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
 
@@ -39,20 +42,20 @@ dependencies {
     testImplementation 'com.h2database:h2'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
 
-    //redis
-    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+//    //redis
+//    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+//
+//    //queryDSL
+//    implementation 'com.querydsl:querydsl-jpa:5.0.0'
+//    annotationProcessor 'com.querydsl:querydsl-apt:5.0.0:jpa'
+//    annotationProcessor 'jakarta.persistence:jakarta.persistence-api'
+//    annotationProcessor 'jakarta.annotation:jakarta.annotation-api'
+//
+//    testImplementation('it.ozimov:embedded-redis:0.7.3'){
+//        exclude group: "org.slf4j", module: "slf4j-simple"
+//    }
 
-    //queryDSL
-    implementation 'com.querydsl:querydsl-jpa:5.0.0'
-    annotationProcessor 'com.querydsl:querydsl-apt:5.0.0:jpa'
-    annotationProcessor 'jakarta.persistence:jakarta.persistence-api'
-    annotationProcessor 'jakarta.annotation:jakarta.annotation-api'
 
-    testImplementation('it.ozimov:embedded-redis:0.7.3'){
-        exclude group: "org.slf4j", module: "slf4j-simple"
-    }
-
-    implementation 'org.apache.commons:commons-lang3:3.12.0'
 }
 ext{
     set('springCloudVersion',"2021.0.1")
