+++ user-api/src/main/java/com/zerobase/cms/user/exception/ExceptionController.java
@@ -8,6 +8,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
+import javax.servlet.ServletException;
+
 @ControllerAdvice
 @Slf4j
 public class ExceptionController {
