+++ user-api/src/main/java/com/zerobase/cms/user/service/customer/CustomerService.java
@@ -3,6 +3,7 @@ package com.zerobase.cms.user.service.customer;
 import com.zerobase.cms.user.domain.model.Customer;
 import com.zerobase.cms.user.domain.repository.CustomerRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -12,16 +13,17 @@ import java.util.Optional;
 public class CustomerService {
 
     private final CustomerRepository customerRepository;
+    private final PasswordEncoder passwordEncoder;
 
     public Optional<Customer> findByIdAndEmail(Long id, String email){
         return customerRepository.findById(id)
                 .stream().filter(customer-> customer.getEmail().equals(email))
                 .findFirst();
     }
-    public Optional<Customer> findValidCustomer(String emaail,String password){
+    public Optional<Customer> findValidCustomer(String email,String password){
 
-        return  customerRepository.findByEmail(emaail).stream().filter(
-                customer -> customer.getPassword().equals(password) && customer.isVerify()
+        return  customerRepository.findByEmail(email).stream().filter(
+                customer -> passwordEncoder.matches(password, customer.getPassword()) && customer.isVerify()
         ).findFirst();
     }
 }
