+++ zerobase-domain/src/test/java/com/zerobase/domain/util/Aes256UtilTest.java
@@ -1,0 +1,16 @@
+package com.zerobase.domain.util;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+class Aes256UtilTest {
+
+    @Test
+    void encrypt(){
+        String encrypt = Aes256Util.encrypt("Hello world");
+        assertEquals(Aes256Util.decrypt(encrypt),"Hello world");
+
+    }
+
+}
